Added Map CSV functionality

+++ retail-be/src/main/java/com/utm/retail_be/entity/ProductCategory.java
@@ -9,7 +9,7 @@ public class ProductCategory {
     @Id
     @Column(name = "product_category_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Short productCategoryID;
+    private Integer productCategoryID;
 
     @Column(nullable = false, length = 255)
     private String name;
